+++ sql/upgrade.sql
@@ -10,4 +10,7 @@ SELECT extversion FROM pg_extension WHERE extname = 'nominatim_fdw';
 ALTER EXTENSION nominatim_fdw UPDATE TO '1.2';
 SELECT extversion FROM pg_extension WHERE extname = 'nominatim_fdw';
 
+/* verify functions are still callable after upgrade */
+SELECT nominatim_fdw_version() IS NOT NULL;
+
 DROP SERVER osm CASCADE;